Alice's DNN
A working Deep Dense Neural Network (Keras). I submitted 3 predictions of 0.70936 (prediction3), 0.72342, and 0.73964 (prediction5).  The prediction4 file got overridden in my directory. It's still on Kaggle but we wouldn't need it anyways since it's not a great score. The increase in performance happened after I increased the number of layers in the network. Also, the validation accuracy/AUC measures are abnormally high and doesn't match our actual submission AUC score (so don't get too excited).